Move get_combinations() helper into tests
